requirements.txt
flask
flask-cors
Flask-SQLAlchemy
mysqlclient